relief_backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
asgiref=3.3.1=pypi_0
astroid=2.4.2=pypi_0
ca-certificates=2020.12.8=h06a4308_0
certifi=2020.12.5=py38h06a4308_0
django=3.1.4=pypi_0
django-cors-headers=3.6.0=pypi_0
django-filter=2.4.0=pypi_0
django-pagination=1.0.7=pypi_0
djangorestframework=3.12.2=pypi_0
isort=5.6.4=pypi_0
lazy-object-proxy=1.4.3=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.3=pypi_0
mccabe=0.6.1=pypi_0
ncurses=6.2=he6710b0_1
openssl=1.1.1i=h27cfd23_0
pip=20.3.3=py38h06a4308_0
pylint=2.6.0=pypi_0
pylint-django=2.3.0=pypi_0
pylint-plugin-utils=0.6=pypi_0
python=3.8.5=h7579374_1
pytz=2020.4=pypi_0
readline=8.0=h7b6447c_0
setuptools=51.0.0=py38h06a4308_2
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
sqlparse=0.4.1=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
